Move common functions to command utils
